smash_core/requirements.txt
# Python dependencies for SMASH Cloud Voice AI
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1
websockets==12.0
python-dotenv==1.0.0
pyaudio==0.2.11
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
speech-recognition==3.10.0
pyttsx3==2.90
elevenlabs==0.2.26
